+++ globus_search_cli/commands/__init__.py
@@ -4,7 +4,6 @@ from globus_search_cli.commands.ingest import ingest_func
 from globus_search_cli.commands.login import login_command
 from globus_search_cli.commands.logout import logout_command
 from globus_search_cli.commands.query import query_func
-from globus_search_cli.commands.query_template import query_template_cmd
 from globus_search_cli.commands.show_index import show_index_cmd
 from globus_search_cli.commands.structured_query import structured_query_func
 from globus_search_cli.commands.subject import subject_cmd
@@ -27,6 +26,5 @@ cli_root.add_command(delete_by_query_cmd)
 cli_root.add_command(subject_cmd)
 cli_root.add_command(entry_cmd)
 cli_root.add_command(task_cmd)
-cli_root.add_command(query_template_cmd)
 cli_root.add_command(login_command)
 cli_root.add_command(logout_command)