+++ globus_search_cli/commands/query_template/__init__.py
@@ -1,3 +1,0 @@
-from globus_search_cli.commands.query_template.main import query_template_cmd
-
-__all__ = ("query_template_cmd",)